refactor(adapter-codex): decompose handle_item_completed into thread_item_render
handle_item_completed was ~117 lines, over the 50-line ceiling, and
event_mapper.rs is already over the 300-line file ceiling. Move the
per-item rendering (and the image-generation/collab helpers it alone
used) into a new sibling module; event_mapper.rs keeps only the
notification-dispatch shell. Pure move, no behavior change.
